backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.19
openai==1.55.3
google-generativeai==0.8.5
httpx==0.27.0
PyPDF2==3.0.1
reportlab==4.0.9
Pillow==11.0.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.3.0

